added the ability to the fourbar to not update at all
TODO: Utilize bulk reads in a seperate thread and bulk writes, read at the everything at the start of a loop in Robot.class, then write all cached operations at the end of the loop all in the main thread and test it in a seperate thread to check for performance.

TODO: create a hz monitoring opmode that gets average, high, low, kalman filtered (smoothed value) for current hz, with the ability to enable and disable subsystems (Camera, Faga, Intake, Lift, Eyelids, Camera, Mogus, Sensor Array, And the Robot wrapper as a whole. Also add some FSM actions to test how they affect the robots hz.